// src/features/drag-flower-to-scene/model/types.ts
// export type DragFlowerPayload = {
//   flowerId: string;
//   title: string;
//   modelLink: string;
// };

// export const DND_MIME = "application/x-flower";


export type DragFlowerPayload = {
  flowerId: string;
  title: string;
  modelLink: string;
};

export const DND_MIME = "application/x-flower";